+++ sdk/model/src/global.d.ts
@@ -1,0 +1,12 @@
+declare global {
+  /** Global factory method returning platforma instance */
+  var getPlatforma: PlatformaFactory;
+  var platforma: Platforma;
+
+  function getEnvironmentValue(name: string): string | undefined;
+
+  /** Global rendering context, present only in rendering environment */
+  var cfgRenderCtx: GlobalCfgRenderCtx;
+}
+
+export {};