+++ sdk/model/src/raw_globals.ts
@@ -1,0 +1,7 @@
+import {} from './global';
+import { getPlatformaInstance } from './internal';
+import { Platforma } from './platforma';
+
+export function getRawPlatformaInstance(): Platforma {
+  return getPlatformaInstance();
+}